% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello}
\usage{
hello(x = c("到期", "付息", "回售权执行", "赎回权执行", "分期还本"))
}
\description{
When the default arguments of the function contain certain non-ASCII characters
on Windows, RStudio IDE will display some malformed info in the console, when
it's trying to display the code completions.
}
\details{
For those non-ASCII characters that won't break the IDE, the info displayed on
Screen is garbage letters.
}
